requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
blinker==1.4
cachelib==0.1.1
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
cloudinary==1.26.0
colorama==0.4.4
distlib==0.3.2
dnspython==2.1.0
email-validator==1.1.2
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.1.0
idna==2.10
install==1.3.4
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.3
jsonify==0.5
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==5.1.0
psycopg2==2.9.1
py3dns==3.2.1
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-http-client==3.3.2
requests==2.25.1
selenium==3.141.0
sendgrid==6.7.1
sib-api-v3-sdk==7.3.0
six==1.15.0
SQLAlchemy==1.3.23
starkbank-ecdsa==1.1.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.7
waitress==2.0.0
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
WTForms==2.3.3
yarl==1.6.3
